+++ box.py
@@ -3,7 +3,7 @@
 #
 # Copyright (c) 2017 - Chris Griffith - MIT License
 """
-Improved dictionary access through recursive dot notation.
+Improved dictionary access through dot notation with additional tools.
 """
 import string
 import sys
@@ -11,9 +11,12 @@ import json
 from uuid import uuid4
 import re
 import collections
+import copy
+from keyword import kwlist
+import warnings
 
 try:
-    from collections.abc import Mapping, Iterable
+    from collections import Iterable, Mapping
 except ImportError:
     Mapping = dict
     Iterable = (tuple, list)
@@ -35,16 +38,13 @@ else:
     from io import open
 
 __all__ = ['Box', 'ConfigBox', 'BoxList', 'SBox',
-           'BoxError']
+           'BoxError', 'BoxKeyError']
 __author__ = 'Chris Griffith'
-__version__ = '3.0.1'
-
-ILLEGAL_ATTRIBUTES = ('if', 'elif', 'else', 'for', 'from', 'as', 'import',
-                      'in', 'not', 'is', 'def', 'class', 'return', 'yield',
-                      'except', 'while', 'raise')
+__version__ = '3.1.0'
 
 BOX_PARAMETERS = ('default_box', 'default_box_attr', 'conversion_box',
-                  'frozen_box', 'camel_killer_box', 'box_it_up')
+                  'frozen_box', 'camel_killer_box', 'box_it_up',
+                  'box_safe_prefix', 'box_duplicates')
 
 _first_cap_re = re.compile('(.)([A-Z][a-z]+)')
 _all_cap_re = re.compile('([a-z0-9])([A-Z])')
@@ -54,7 +54,11 @@ class BoxError(Exception):
     """Non standard dictionary exceptions"""
 
 
+class BoxKeyError(BoxError, KeyError, AttributeError):
+    """Key does not exist"""
+
 # Abstract converter functions for use in any Box class
+
 
 def _to_json(obj, filename=None,
              encoding="utf-8", errors="strict", **json_kwargs):
@@ -69,10 +73,14 @@ def _to_json(obj, filename=None,
 
 
 def _from_json(json_string=None, filename=None,
-               encoding="utf-8", errors="strict", **kwargs):
+               encoding="utf-8", errors="strict", multiline=False, **kwargs):
     if filename:
         with open(filename, 'r', encoding=encoding, errors=errors) as f:
-            data = json.load(f, **kwargs)
+            if multiline:
+                data = [json.loads(line.strip(), **kwargs) for line in f
+                        if line.strip() and not line.strip().startswith("#")]
+            else:
+                data = json.load(f, **kwargs)
     elif json_string:
         data = json.loads(json_string, **kwargs)
     else:
@@ -115,10 +123,10 @@ def _safe_key(key):
     try:
         return str(key)
     except UnicodeEncodeError:
-        return key.encode(encoding="utf-8", errors="ignore")
+        return key.encode("utf-8", "ignore")
 
 
-def _safe_attr(attr, camel_killer=False):
+def _safe_attr(attr, camel_killer=False, replacement_char='x'):
     """Convert a key into something that is accessible as an attribute"""
     allowed = string.ascii_letters + string.digits + '_'
 
@@ -139,10 +147,10 @@ def _safe_attr(attr, camel_killer=False):
     except (ValueError, IndexError):
         pass
     else:
-        out = 'x{0}'.format(out)
+        out = '{0}{1}'.format(replacement_char, out)
 
-    if out in ILLEGAL_ATTRIBUTES:
-        out = 'x{0}'.format(out)
+    if out in kwlist:
+        out = '{0}{1}'.format(replacement_char, out)
 
     return re.sub('_+', '_', out)
 
@@ -156,7 +164,7 @@ def _camel_killer(attr):
     try:
         attr = str(attr)
     except UnicodeEncodeError:
-        attr = attr.encode(encoding="utf-8", errors="ignore")
+        attr = attr.encode("utf-8", "ignore")
 
     s1 = _first_cap_re.sub(r'\1_\2', attr)
     s2 = _all_cap_re.sub(r'\1_\2', s1)
@@ -177,6 +185,68 @@ def _recursive_tuples(iterable, box_class, recreate_tuples=False, **kwargs):
     return tuple(out_list)
 
 
+def _conversion_checks(item, keys, box_config, check_only=False,
+                       pre_check=False):
+    """
+    Internal use for checking if a duplicate safe attribute already exists
+
+    :param item: Item to see if a dup exists
+    :param keys: Keys to check against
+    :param box_config: Easier to pass in than ask for specfic items
+    :param check_only: Don't bother doing the conversion work
+    :param pre_check: Need to add the item to the list of keys to check
+    :return: the original unmodified key, if exists and not check_only
+    """
+    if box_config['box_duplicates'] != 'ignore':
+        if pre_check:
+            keys = list(keys) + [item]
+
+        key_list = [(k,
+                     _safe_attr(k, camel_killer=box_config['camel_killer_box'],
+                                replacement_char=box_config['box_safe_prefix']
+                                )) for k in keys]
+        if len(key_list) > len(set(x[1] for x in key_list)):
+            seen = set()
+            dups = set()
+            for x in key_list:
+                if x[1] in seen:
+                    dups.add("{0}({1})".format(x[0], x[1]))
+                seen.add(x[1])
+            if box_config['box_duplicates'].startswith("warn"):
+                warnings.warn('Duplicate conversion attributes exist: '
+                              '{0}'.format(dups))
+            else:
+                raise BoxError('Duplicate conversion attributes exist: '
+                               '{0}'.format(dups))
+    if check_only:
+        return
+    # This way will be slower for warnings, as it will have double work
+    # But faster for the default 'ignore'
+    for k in keys:
+        if item == _safe_attr(k, camel_killer=box_config['camel_killer_box'],
+                              replacement_char=box_config['box_safe_prefix']):
+            return k
+
+
+def _get_box_config(cls, kwargs):
+    return {
+            # Internal use only
+            '__converted': set(),
+            '__box_heritage': kwargs.pop('__box_heritage', None),
+            '__hash': None,
+            '__created': False,
+            # Can be changed by user after box creation
+            'default_box': kwargs.pop('default_box', False),
+            'default_box_attr': kwargs.pop('default_box_attr', cls),
+            'conversion_box': kwargs.pop('conversion_box', True),
+            'box_safe_prefix': kwargs.pop('box_safe_prefix', 'x'),
+            'frozen_box': kwargs.pop('frozen_box', False),
+            'camel_killer_box': kwargs.pop('camel_killer_box', False),
+            'modify_tuples_box': kwargs.pop('modify_tuples_box', False),
+            'box_duplicates': kwargs.pop('box_duplicates', 'ignore')
+            }
+
+
 class Box(dict):
     """
 
@@ -192,26 +262,28 @@ class Box(dict):
     :param conversion_box: Check for near matching keys as attributes
     :param modify_tuples_box: Recreate incoming tuples with dicts into Boxes
     :param box_it_up: Recursively create all Boxes from the start
+    :param box_safe_prefix: Conversion box prefix for unsafe attributes
+    :param box_duplicates: "ignore", "error" or "warn" when duplicates exists
+        in a conversion_box
     """
 
     _protected_keys = dir({}) + ['to_dict', 'tree_view', 'to_json', 'to_yaml',
                                  'from_yaml', 'from_json']
 
+    def __new__(cls, *args, **kwargs):
+        """
+        Due to the way pickling works in python 3, we need to make sure
+        the box config is created as early as possible.
+        """
+        obj = super(Box, cls).__new__(cls, *args, **kwargs)
+        obj._box_config = _get_box_config(cls, kwargs)
+        return obj
+
     def __init__(self, *args, **kwargs):
-        self._box_config = {
-            # Internal use only
-            '__converted': set(),
-            '__box_heritage': kwargs.pop('__box_heritage', None),
-            '__hash': None,
-            '__created': False,
-            # Can be changed by user after box creation
-            'default_box': kwargs.pop('default_box', False),
-            'default_box_attr': kwargs.pop('default_box_attr', self.__class__),
-            'conversion_box': kwargs.pop('conversion_box', True),
-            'frozen_box': kwargs.pop('frozen_box', False),
-            'camel_killer_box': kwargs.pop('camel_killer_box', False),
-            'modify_tuples_box': kwargs.pop('modify_tuples_box', False)
-            }
+        self._box_config = _get_box_config(self.__class__, kwargs)
+        if (not self._box_config['conversion_box'] and
+           self._box_config['box_duplicates'] != "ignore"):
+            raise BoxError('box_duplicates are only for conversion_boxes')
         if len(args) == 1:
             if isinstance(args[0], basestring):
                 raise ValueError('Cannot extrapolate Box from string')
@@ -239,7 +311,8 @@ class Box(dict):
                 v = self
             self[k] = v
 
-        if self._box_config['frozen_box'] or box_it:
+        if (self._box_config['frozen_box'] or box_it or
+           self._box_config['box_duplicates'] != 'ignore'):
             self.box_it_up()
 
         self._box_config['__created'] = True
@@ -251,7 +324,9 @@ class Box(dict):
         any of those sub box objects.
         """
         for k in self:
-            if hasattr(self[k], 'box_it_up'):
+            _conversion_checks(k, self.keys(), self._box_config,
+                               check_only=True)
+            if self[k] is not self and hasattr(self[k], 'box_it_up'):
                 self[k].box_it_up()
 
     def __hash__(self):
@@ -273,7 +348,7 @@ class Box(dict):
         for key in self.keys():
             key = _safe_key(key)
             if (' ' not in key and key[0] not in string.digits and
-                    key not in ILLEGAL_ATTRIBUTES):
+                    key not in kwlist):
                 for letter in key:
                     if letter not in allowed:
                         break
@@ -284,7 +359,9 @@ class Box(dict):
             key = _safe_key(key)
             if key not in items:
                 if self._box_config['conversion_box']:
-                    key = _safe_attr(key, camel_killer=kill_camel)
+                    key = _safe_attr(key, camel_killer=kill_camel,
+                                     replacement_char=self._box_config[
+                                         'box_safe_prefix'])
                     if key:
                         items.add(key)
             if kill_camel:
@@ -302,27 +379,57 @@ class Box(dict):
     def __contains__(self, item):
         return dict.__contains__(self, item) or hasattr(self, item)
 
+    def copy(self):
+        return self.__class__(super(self.__class__, self).copy())
+
+    def __copy__(self):
+        return self.__class__(super(self.__class__, self).copy())
+
+    def __deepcopy__(self, memodict=None):
+        out = self.__class__()
+        memodict = memodict or {}
+        memodict[id(self)] = out
+        for k, v in self.items():
+            out[copy.deepcopy(k, memodict)] = copy.deepcopy(v, memodict)
+        return out
+
+    def __setstate__(self, state):
+        self._box_config = state['_box_config']
+        self.__dict__.update(state)
+
     def __getitem__(self, item):
         try:
             value = super(Box, self).__getitem__(item)
         except KeyError as err:
+            if item == '_box_config':
+                raise BoxError('_box_config key must exist and does not. '
+                               'This is most likely a bug, please report.')
             default_value = self._box_config['default_box_attr']
             if self._box_config['default_box']:
-                if isinstance(default_value, collections.Callable):
-                    if default_value.__name__ == 'Box':
-                        return self.__class__(__box_heritage=(self, item),
-                                              **self.__box_config())
+                if default_value is self.__class__:
+                    return self.__class__(__box_heritage=(self, item),
+                                          **self.__box_config())
+                elif isinstance(default_value, collections.Callable):
                     return default_value()
                 elif hasattr(default_value, 'copy'):
                     return default_value.copy()
                 return default_value
-            raise err
+            raise BoxKeyError(str(err))
         else:
             return self.__convert_and_store(item, value)
 
+    def values(self):
+        return [self[x] for x in self]
+
+    def items(self):
+        return [(x, self[x]) for x in self]
+
     def __box_config(self):
-        return {k: v for k, v in self._box_config.items()
-                if not k.startswith("__")}
+        out = {}
+        for k, v in self._box_config.copy().items():
+            if not k.startswith("__"):
+                out[k] = v
+        return out
 
     def __convert_and_store(self, item, value):
         if item in self._box_config['__converted']:
@@ -373,16 +480,16 @@ class Box(dict):
             except KeyError:
                 if item == '_box_config':
                     raise BoxError('_box_config key must exist')
-                kill_camel = self._box_config.get('camel_killer_box', False)
-                if self._box_config.get('conversion_box', False) and item:
-                    for k in self.keys():
-                        if item == _safe_attr(k, camel_killer=kill_camel):
-                            return self.__getitem__(k)
+                kill_camel = self._box_config['camel_killer_box']
+                if self._box_config['conversion_box'] and item:
+                    k = _conversion_checks(item, self.keys(), self._box_config)
+                    if k:
+                        return self.__getitem__(k)
                 if kill_camel:
                     for k in self.keys():
                         if item == _camel_killer(k):
                             return self.__getitem__(k)
-            raise err
+            raise BoxKeyError(str(err))
         else:
             if item == '_box_config':
                 return value
@@ -392,6 +499,9 @@ class Box(dict):
         if (key != '_box_config' and self._box_config['__created'] and
                 self._box_config['frozen_box']):
             raise BoxError('Box is frozen')
+        if self._box_config['conversion_box']:
+            _conversion_checks(key, self.keys(), self._box_config,
+                               check_only=True, pre_check=True)
         super(Box, self).__setitem__(key, value)
         self.__create_lineage()
 
@@ -409,14 +519,12 @@ class Box(dict):
             if (key not in self.keys() and
                     (self._box_config['conversion_box'] or
                      self._box_config['camel_killer_box'])):
-                for each_key in self:
-                    if self._box_config['conversion_box']:
-                        if key == _safe_attr(
-                                each_key,
-                                self._box_config['camel_killer_box']):
-                            self[each_key] = value
-                            break
-                    elif self._box_config['camel_killer_box']:
+                if self._box_config['conversion_box']:
+                    k = _conversion_checks(key, self.keys(),
+                                           self._box_config)
+                    self[key if not k else k] = value
+                elif self._box_config['camel_killer_box']:
+                    for each_key in self:
                         if key == _camel_killer(each_key):
                             self[each_key] = value
                             break
@@ -460,7 +568,6 @@ class Box(dict):
 
         :return: python dictionary of this Box
         """
-
         out_dict = dict(self)
         for k, v in out_dict.items():
             if v is self:
@@ -517,7 +624,7 @@ class Box(dict):
                   encoding="utf-8", errors="strict", **kwargs):
         """
         Transform a json object string into a Box object. If the incoming
-        json is a list, you must use BoxList.from_json. 
+        json is a list, you must use BoxList.from_json.
 
         :param json_string: string to pass to `json.loads`
         :param filename: filename to open and pass to `json.load`
@@ -618,11 +725,32 @@ class BoxList(list):
                         BoxList(p_object))
         super(BoxList, self).insert(index, p_object)
 
+    def __getstate__(self):
+        return {'box_class': self.box_class,
+                'box_options': self.box_options,
+                'box_org_ref': self.box_org_ref}
+
+    def __setstate__(self, state):
+        self.box_class = state['box_class']
+        self.box_options = state['box_options']
+        self.box_org_ref = state['box_org_ref']
+
     def __repr__(self):
         return "<BoxList: {0}>".format(self.to_list())
 
     def __str__(self):
         return str(self.to_list())
+
+    def __copy__(self):
+        return BoxList(x for x in self)
+
+    def __deepcopy__(self, memodict=None):
+        out = self.__class__()
+        memodict = memodict or {}
+        memodict[id(self)] = out
+        for k in self:
+            out.append(copy.deepcopy(k))
+        return out
 
     def to_list(self):
         new_list = []
@@ -638,22 +766,31 @@ class BoxList(list):
         return new_list
 
     def to_json(self, filename=None,
-                encoding="utf-8", errors="strict", **json_kwargs):
+                encoding="utf-8", errors="strict",
+                multiline=False, **json_kwargs):
         """
         Transform the BoxList object into a JSON string.
 
         :param filename: If provided will save to file
         :param encoding: File encoding
         :param errors: How to handle encoding errors
+        :param multiline: Put each item in list onto it's own line
         :param json_kwargs: additional arguments to pass to json.dump(s)
         :return: string of JSON or return of `json.dump`
         """
-        return _to_json(self.to_list(), filename=filename,
-                        encoding=encoding, errors=errors, **json_kwargs)
+        if filename and multiline:
+            lines = [_to_json(item, filename=False, encoding=encoding,
+                              errors=errors, **json_kwargs) for item in self]
+            with open(filename, 'w', encoding=encoding, errors=errors) as f:
+                f.write("\n".join(lines).decode('utf-8') if
+                        sys.version_info < (3, 0) else "\n".join(lines))
+        else:
+            return _to_json(self.to_list(), filename=filename,
+                            encoding=encoding, errors=errors, **json_kwargs)
 
     @classmethod
-    def from_json(cls, json_string=None, filename=None,
-                  encoding="utf-8", errors="strict", **kwargs):
+    def from_json(cls, json_string=None, filename=None, encoding="utf-8",
+                  errors="strict", multiline=False, **kwargs):
         """
         Transform a json object string into a BoxList object. If the incoming
         json is a dict, you must use Box.from_json.
@@ -662,6 +799,7 @@ class BoxList(list):
         :param filename: filename to open and pass to `json.load`
         :param encoding: File encoding
         :param errors: How to handle encoding errors
+        :param multiline: One object per line
         :param kwargs: parameters to pass to `Box()` or `json.loads`
         :return: BoxList object from json data
         """
@@ -670,8 +808,8 @@ class BoxList(list):
             if arg in BOX_PARAMETERS:
                 bx_args[arg] = kwargs.pop(arg)
 
-        data = _from_json(json_string, filename=filename,
-                          encoding=encoding, errors=errors, **kwargs)
+        data = _from_json(json_string, filename=filename, encoding=encoding,
+                          errors=errors, multiline=multiline, **kwargs)
 
         if not isinstance(data, list):
             raise BoxError('json data not returned as a list, '
@@ -724,7 +862,7 @@ class BoxList(list):
 
     def box_it_up(self):
         for v in self:
-            if hasattr(v, 'box_it_up'):
+            if hasattr(v, 'box_it_up') and v is not self:
                 v.box_it_up()